backend/requirements.txt
aioredis==1.1.0
amqp==2.3.2
asgiref==2.3.2
async-timeout==3.0.0
asyncio==3.4.3
attrs==18.1.0
autobahn==18.8.1
Automat==0.7.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
channels==2.1.3
channels-redis==2.3.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==2.2.2
Django==2.1
django-celery-results==1.0.1
django-cors-headers==2.4.0
django-rest-swagger==2.2.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
gunicorn==19.9.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
msgpack==0.5.6
openapi-codec==1.3.2
Pillow==5.2.0
psycopg2-binary==2.7.5
PyHamcrest==1.9.0
PyJWT==1.6.4
pytz==2018.5
requests==2.19.1
simplejson==3.16.0
six==1.11.0
Twisted==18.7.0
txaio==18.7.1
uritemplate==3.0.0
urllib3==1.23
vine==1.1.4
zope.interface==4.5.0
